pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "setuptools-scm >= 8"]
build-backend = "setuptools.build_meta"

[project]
name = "coh2_live_stats"
dynamic = ["version"]
dependencies = [
    "colorama >= 0.4.6",
    "httpx >= 0.27.0",
    "prettytable >= 3.10.0",
    "pydantic-settings >= 2.2.1",
    "watchdog >= 4.0.0",
]
requires-python = ">= 3.12"
authors = [
    { name = "Andreas Becker", email = "brofi.archlinux@gmail.com" },
]
description = "Show player and team stats of a currently played, replayed or last played Company of Heroes 2 match."
readme = "README.md"
license = { file = "COPYING" }
keywords = ["CoH2", "Company of Heroes 2"]
classifiers = [
    "Environment :: Console",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Operating System :: Microsoft :: Windows",
    "Topic :: Games/Entertainment :: Real Time Strategy"
]

# No better solution for this yet (https://peps.python.org/pep-0735/)
[project.optional-dependencies]
dev = [
    "ruff >= 0.4.1",
    "pre-commit >= 3.7.0",
]
build = [
    "build >= 1.2.1",
    "invoke >= 2.2.0",
    "pyinstaller >= 6.6.0",
    "pytest >= 8.1.1",
    "pytest-asyncio >= 0.23.6",
    "respx >= 0.21.1",
    "tomlkit >= 0.12.4",
]

[project.scripts]
coh2livestats = "coh2_live_stats.__main__:run"

[tool.setuptools.package-data]
coh2_live_stats = ["*.toml"]
"coh2_live_stats.res" = ["*.wav"]

[tool.setuptools_scm]
version_file = "src/coh2_live_stats/version.py"

[tool.mypy]
plugins = ["pydantic.mypy"]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
extra_checks = true
check_untyped_defs = true
disallow_subclassing_any = true
disallow_untyped_decorators = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_reexport = true
warn_return_any = true

[[tool.mypy.overrides]]
module = "invoke"
no_implicit_reexport = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff.lint]
preview = true
select = [
    "F", # Pyflakes
    "E", "W", # pycodestyle
    "C90", # mccabe
    "I", # isort
    "N", # pep8-naming
    "D", # pydocstyle
    "UP", # pyupgrade
    "YTT", # flake8-2020
    "ANN", # flake8-annotations
    "ASYNC", # flake8-async
    "TRIO", # flake8-trio
    "S", # flake8-bandit
    "BLE", # flake8-blind-except
    "FBT", # flake8-boolean-trap
    "B", # flake8-bugbear
    "A", # flake8-builtins
    "COM", # flake8-commas
    "CPY", # flake8-copyright # preview
    "C4", # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "EM", # flake8-errmsg
    "EXE", # flake8-executable
    "FA", # flake8-future-annotations
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "LOG", # flake8-logging
    "G", # flake8-logging-format
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT", # flake8-pytest-style
    "Q004", # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SLF", # flake8-self
    "SLOT", # flake8-slots
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "TCH", # flake8-type-checking
    "INT", # flake8-gettext
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "TD", # flake8-todos
    "PGH", # pygrep-hooks
    "PL", # Pylint
    "TRY", # tryceratops
    "FLY", # flynt
    "PERF", # Perflint
    "FURB", # refurb # preview
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501", # fomatter sometimes allows code to exceed 'line-length'
    "TD001", "TD002", "TD003", # no invalid tag and missing author or issue link
    "TRY003", # too restrictive
    "D203", # use D211 instead
    "D213", # use D212 instead
    # conflicts with formatter
    "W191", "E111", "E114", "E117",
    "D206", "D300",
    "COM812", "COM819",
    "ISC001", "ISC002"
]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.isort]
split-on-trailing-comma = false
classes = ["URL"]

[tool.ruff.format]
docstring-code-format = true
quote-style = "single"
skip-magic-trailing-comma = true
